+++ src/main/java/cn/nukkit/network/protocol/InteractPacket.java
@@ -8,12 +8,12 @@ public class InteractPacket extends DataPacket {
     public static final byte NETWORK_ID = Info.INTERACT_PACKET;
 
     public long eid;
-    public int action;
+    public byte action;
     public long target;
 
     @Override
     public void decode() {
-        action = getByte();
+        action = (byte) getByte();
         target = getLong();
     }
