+++ src/main/java/cn/nukkit/network/protocol/MobEquipmentPacket.java
@@ -32,7 +32,7 @@ public class MobEquipmentPacket extends DataPacket {
         this.reset();
         this.putLong(this.eid);
         this.putSlot(this.item);
-        this.putByte(this.slot);
-        this.putByte(this.selectedSlot);
+        this.putByte((byte) this.slot);
+        this.putByte((byte) this.selectedSlot);
     }
 }
